Silent gcc compiler warnings. - fix type mismatches - clean up unused parameters (except for BLAS.hpp)

+++ src/Main/InputOutput.cpp
@@ -12,7 +12,7 @@ void readMainInputFile(const char* FileName)
 {
   
   //std::cout << "Reading main input file: " << FileName << std::endl;
-  int length = strlen(FileName);
+  size_t length = strlen(FileName);
   simInfo.MCInputFile = new char[length+1]();
   strncpy(simInfo.MCInputFile, FileName, length+1);
   //std::cout << "simInfo.MCInputFile set to be: " << simInfo.MCInputFile << std::endl;
